Update README.md
How do I interpret user needs and implement them into a program? How does creating “user stories” help with this?
To interpret user needs one can starts with system analysis. It can be completed by asking for feedback from old users and covering the alteration in the code. The specific inputs and outputs for the system should be identified. User stories are great way to implement user needs. This tool mainly used in Agile development. It helps with the understanding of the end user’s perspective. In addition, user stories help to determine what the user wants, and help create a simple requirement description of their wants. 
How do I approach developing programs? What agile processes do I hope to incorporate into my future development work?
There are different ways to approach developing programs. For example, by using different project management strategies. Waterfall methodology is one of them. It helps in projecting the needs and adopting certain strategies. Some other approaches are prototype, agile, iterative, and incremental. 

What does it mean to be a good team member in software development?

A good team member in software development means to follow the team standards and project styles in coding and designing process. It is vital to maintain a sense of progress in the team.  Every aspect of the progress must be considered. One should always react in a friendly manner and be accountable for their responsibility in regard to the project.